build(typescript): Phase 5 batch 5 — convert tests/ .mjs to .ts (25 files)
Converts tests/chain-registrations.test.mjs through
tests/concentration.test.mjs (25 files) to TypeScript as part of the
tests/ migration (474 files total, batched).
