+++ tests/chutes-call-subnet-surface-verify.test.ts
@@ -1,46 +1,38 @@
-// SN29 (Coldint) end-to-end verification for the call_subnet_surface MCP
-// tool (metagraphed#7045, MCP execute Phase 1 follow-up #7014/#7215). Unlike
-// tests/call-subnet-surface-mcp.test.mjs -- which proves the tool wiring
-// with synthetic surfaces -- this file pins SN29's registry surface
-// (registry/subnets/coldint.json) to the tool's contract, so a future edit
-// that regresses its callability (flipping to HEAD, marking it
-// auth_required, disabling its probe) is caught here.
+// SN64 (SN64) end-to-end verification for the call_subnet_surface MCP tool
+// (metagraphed#7077, MCP execute Phase 1 follow-up #7014/#7215). Unlike
+// tests/call-subnet-surface-mcp.test.mjs -- which proves the tool wiring with
+// synthetic surfaces -- this file pins SN64's *real* registry surface config
+// (registry/subnets/chutes.json) to the tool's contract, so a future edit that
+// regresses its callability (flipping to HEAD, marking it auth_required,
+// disabling its probe) is caught here.
 //
-// The surface is the public no-auth competitions-config API
-// (sn-29-coldint-competitions-api, GET
-// https://github.com/coldint/sn29/raw/main/competitions.json, JSON, single
-// fixed endpoint -- no schema). Live-verified 2026-07-21: the `raw` URL
-// 302-redirects to raw.githubusercontent.com (which fetch() follows
-// transparently, same as callSubnetSurface's own manual redirect handling),
-// resolving to HTTP 200 application/json. Cross-checked directly against
-// the GitHub contents API (a separate, authoritative source): the file is
-// genuinely only 3 bytes right now -- `{}` -- not a fetch or caching
-// artifact. That's live upstream data (no active competitions configured
-// at verification time), not a probe-config defect; the surface's kind,
-// auth, and probe shape all correctly match how it actually behaves. The
-// fixture below mirrors that live (redirect-resolved) response rather than
-// fetching it, keeping the test hermetic while still exercising the JSON
-// parse-and-return path.
+// The surface is the public no-auth SN64 API health endpoint
+// (sn-64-chutes-health, GET https://chutes.ai/api/health, JSON, single fixed endpoint -- no schema). Live-verified
+// 2026-07-21 to return HTTP 200 application/json {"status":"healthy"}. The
+// fixture below mirrors that live response rather than fetching it, keeping the
+// test hermetic while still exercising the JSON parse-and-return path.
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { describe, test } from "vitest";
 import { callSubnetSurface } from "../src/call-subnet-surface.ts";
 import { handleMcpRequest } from "../src/mcp-server.mjs";
+import type { Row } from "./row-type.ts";
 
-const SURFACE_ID = "sn-29-coldint-competitions-api";
+const SURFACE_ID = "sn-64-chutes-health";
 
-const registry = JSON.parse(
+const registry: Row = JSON.parse(
   readFileSync(
-    fileURLToPath(new URL("../registry/subnets/coldint.json", import.meta.url)),
+    fileURLToPath(new URL("../registry/subnets/chutes.json", import.meta.url)),
     "utf8",
   ),
 );
-const SURFACE = registry.surfaces.find((surface) => surface.id === SURFACE_ID);
+const SURFACE = registry.surfaces.find(
+  (surface: Row) => surface.id === SURFACE_ID,
+);
 
-// The live (redirect-resolved) response body: no active competitions
-// configured at verification time.
-const BODY = {};
+// A faithful subset of the live https://chutes.ai/api/health response body.
+const BODY = { status: "healthy" };
 
 function upstreamResponse() {
   return new Response(JSON.stringify(BODY), {
@@ -49,31 +41,30 @@ function upstreamResponse() {
   });
 }
 
-describe("SN29 Coldint call_subnet_surface verification (#7045)", () => {
+describe("SN64 SN64 call_subnet_surface verification (#7077)", () => {
   test("the registry surface exists and is configured to be callable", () => {
     assert.ok(SURFACE, `registry surface ${SURFACE_ID} is present`);
     assert.equal(SURFACE.kind, "subnet-api");
     assert.equal(SURFACE.auth_required, false);
     assert.equal(SURFACE.probe?.enabled, true);
+    // No-auth GET returning JSON.
     assert.equal(SURFACE.probe?.method, "GET");
     assert.equal(SURFACE.probe?.expect, "json");
-    assert.equal(
-      SURFACE.url,
-      "https://github.com/coldint/sn29/raw/main/competitions.json",
-    );
+    assert.equal(SURFACE.url, "https://chutes.ai/api/health");
+    // Single fixed endpoint -- no machine-readable schema is expected.
     assert.equal(SURFACE.schema_url, undefined);
   });
 
   test("callSubnetSurface returns the real JSON body using the surface's own url + GET", async () => {
-    let requestedUrl;
-    let requestedMethod;
+    let requestedUrl: string | undefined;
+    let requestedMethod: string | undefined;
     const result = await callSubnetSurface(SURFACE, {
       isUnsafeUrl: async () => false,
-      fetchImpl: async (url, init) => {
+      fetchImpl: (async (url: string | URL, init?: RequestInit) => {
         requestedUrl = String(url);
-        requestedMethod = init.method;
+        requestedMethod = init?.method;
         return upstreamResponse();
-      },
+      }) as typeof fetch,
     });
     assert.equal(result.ok, true);
     assert.equal(requestedUrl, SURFACE.url);
@@ -81,21 +72,21 @@ describe("SN29 Coldint call_subnet_surface verification (#7045)", () => {
     assert.equal(result.status_code, 200);
     assert.equal(result.content_type, "application/json");
     assert.equal(result.truncated, false);
-    assert.deepEqual(result.body, BODY);
+    assert.equal((result.body as Row).status, "healthy");
   });
 
   test("end-to-end through the call_subnet_surface MCP tool, resolved by surface id", async () => {
     const catalog = {
-      surfaces: [{ ...SURFACE, surface_id: SURFACE.id, netuid: 29 }],
+      surfaces: [{ ...SURFACE, surface_id: SURFACE.id, netuid: 64 }],
     };
     const deps = {
-      readArtifact: async (_env, path) =>
+      readArtifact: async (_env: Row, path: string) =>
         path === "/metagraph/operational-surfaces.json"
           ? { ok: true, data: catalog }
           : { ok: false, status: 404 },
     };
     const originalFetch = globalThis.fetch;
-    globalThis.fetch = async (input) => {
+    globalThis.fetch = (async (input: string | URL | Request) => {
       const url = String(input);
       if (url.startsWith("https://cloudflare-dns.com/dns-query")) {
         return new Response(JSON.stringify({ Status: 0 }), {
@@ -103,7 +94,7 @@ describe("SN29 Coldint call_subnet_surface verification (#7045)", () => {
         });
       }
       return upstreamResponse();
-    };
+    }) as typeof fetch;
     try {
       const response = await handleMcpRequest(
         new Request("https://metagraph.sh/mcp", {
@@ -122,11 +113,11 @@ describe("SN29 Coldint call_subnet_surface verification (#7045)", () => {
         {},
         deps,
       );
-      const result = (await response.json()).result;
+      const result = ((await response.json()) as Row).result;
       assert.equal(result.isError, false);
       assert.equal(result.structuredContent.surface_id, SURFACE_ID);
       assert.equal(result.structuredContent.status_code, 200);
-      assert.deepEqual(result.structuredContent.body, BODY);
+      assert.equal((result.structuredContent.body as Row).status, "healthy");
     } finally {
       globalThis.fetch = originalFetch;
     }
