+++ tests/compute-horde-call-subnet-surface-verify.test.ts
@@ -1,46 +1,56 @@
-// SN83 (CliqueAI) end-to-end verification for the call_subnet_surface MCP tool
-// (metagraphed#7058, MCP execute Phase 1 follow-up #7014/#7215; issue #7096).
-// Unlike tests/call-subnet-surface-mcp.test.mjs -- which proves the tool wiring
-// with synthetic surfaces -- this file pins SN83's *real* registry surface
-// config (registry/subnets/cliqueai.json) to the tool's contract, so a future
+// SN12 (Compute Horde) end-to-end verification for the call_subnet_surface
+// MCP tool (metagraphed#7028, MCP execute Phase 1 follow-up #7014/#7215).
+// Unlike tests/call-subnet-surface-mcp.test.mjs -- which proves the tool
+// wiring with synthetic surfaces -- this file pins SN12's registry surface
+// (registry/subnets/compute-horde.json) to the tool's contract, so a future
 // edit that regresses its callability (flipping to HEAD, marking it
-// auth_required, disabling its probe, moving the url) is caught here.
+// auth_required, disabling its probe) is caught here.
 //
-// The surface is the public no-auth SN83 indexed subnet feed on TaoMarketCap
-// (sn-83-taomarketcap-subnet-api, GET
-// https://api.taomarketcap.com/public/v1/subnets/83/, JSON, single fixed
-// endpoint -- no schema). Live-verified 2026-07-21 to return HTTP 200
-// application/json -- an object self-reporting netuid 83 (id, is_active,
-// mechanism_count, latest_snapshot). The fixture below mirrors that live
-// response rather than fetching it, keeping the test hermetic while still
-// exercising the tool's JSON parse-and-return path.
+// The surface is the public no-auth Grafana OpenAPI schema
+// (sn-12-compute-horde-grafana-openapi, GET
+// https://grafana.bittensor.church/public/openapi3.json, JSON, has captured
+// schema). Live-verified 2026-07-21 to return HTTP 200 application/json (a
+// ~735 KB OpenAPI 3 document). The fixture below mirrors the shape of that
+// live response with a minimal document rather than fetching it, keeping
+// the test hermetic while still exercising the JSON parse-and-return path.
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { describe, test } from "vitest";
 import { callSubnetSurface } from "../src/call-subnet-surface.ts";
 import { handleMcpRequest } from "../src/mcp-server.mjs";
+import type { Row } from "./row-type.ts";
 
-const SURFACE_ID = "sn-83-taomarketcap-subnet-api";
-const NETUID = 83;
+const SURFACE_ID = "sn-12-compute-horde-grafana-openapi";
 
-const registry = JSON.parse(
+const registry: Row = JSON.parse(
   readFileSync(
     fileURLToPath(
-      new URL("../registry/subnets/cliqueai.json", import.meta.url),
+      new URL("../registry/subnets/compute-horde.json", import.meta.url),
     ),
     "utf8",
   ),
 );
-const SURFACE = registry.surfaces.find((surface) => surface.id === SURFACE_ID);
+const SURFACE = registry.surfaces.find(
+  (surface: Row) => surface.id === SURFACE_ID,
+);
 
-// A faithful subset of the live SN83 TaoMarketCap subnet feed.
+// A faithful subset of the live Grafana OpenAPI document's shape.
 const BODY = {
-  id: "83",
-  netuid: 83,
-  is_active: true,
-  mechanism_count: 1,
-  latest_snapshot_id: "8667931-83",
+  components: {
+    responses: {
+      GetAllIntervalsResponse: {
+        content: {
+          "application/json": {
+            schema: {
+              type: "array",
+              items: { $ref: "#/components/schemas/GettableTimeIntervals" },
+            },
+          },
+        },
+      },
+    },
+  },
 };
 
 function upstreamResponse() {
@@ -50,33 +60,31 @@ function upstreamResponse() {
   });
 }
 
-describe("SN83 CliqueAI call_subnet_surface verification (#7058)", () => {
+describe("SN12 Compute Horde call_subnet_surface verification (#7028)", () => {
   test("the registry surface exists and is configured to be callable", () => {
     assert.ok(SURFACE, `registry surface ${SURFACE_ID} is present`);
-    assert.equal(SURFACE.kind, "subnet-api");
+    assert.equal(SURFACE.kind, "openapi");
     assert.equal(SURFACE.auth_required, false);
     assert.equal(SURFACE.probe?.enabled, true);
-    // No-auth GET returning JSON.
     assert.equal(SURFACE.probe?.method, "GET");
     assert.equal(SURFACE.probe?.expect, "json");
     assert.equal(
       SURFACE.url,
-      "https://api.taomarketcap.com/public/v1/subnets/83/",
+      "https://grafana.bittensor.church/public/openapi3.json",
     );
-    // Single fixed endpoint -- no machine-readable schema is expected.
-    assert.equal(SURFACE.schema_url, undefined);
+    assert.equal(SURFACE.schema_url, SURFACE.url);
   });
 
   test("callSubnetSurface returns the real JSON body using the surface's own url + GET", async () => {
-    let requestedUrl;
-    let requestedMethod;
+    let requestedUrl: string | undefined;
+    let requestedMethod: string | undefined;
     const result = await callSubnetSurface(SURFACE, {
       isUnsafeUrl: async () => false,
-      fetchImpl: async (url, init) => {
+      fetchImpl: (async (url: string | URL, init?: RequestInit) => {
         requestedUrl = String(url);
-        requestedMethod = init.method;
+        requestedMethod = init?.method;
         return upstreamResponse();
-      },
+      }) as typeof fetch,
     });
     assert.equal(result.ok, true);
     assert.equal(requestedUrl, SURFACE.url);
@@ -84,22 +92,21 @@ describe("SN83 CliqueAI call_subnet_surface verification (#7058)", () => {
     assert.equal(result.status_code, 200);
     assert.equal(result.content_type, "application/json");
     assert.equal(result.truncated, false);
-    assert.equal(result.body.netuid, NETUID);
-    assert.equal(typeof result.body.is_active, "boolean");
+    assert.deepEqual(result.body, BODY);
   });
 
   test("end-to-end through the call_subnet_surface MCP tool, resolved by surface id", async () => {
     const catalog = {
-      surfaces: [{ ...SURFACE, surface_id: SURFACE.id, netuid: NETUID }],
+      surfaces: [{ ...SURFACE, surface_id: SURFACE.id, netuid: 12 }],
     };
     const deps = {
-      readArtifact: async (_env, path) =>
+      readArtifact: async (_env: Row, path: string) =>
         path === "/metagraph/operational-surfaces.json"
           ? { ok: true, data: catalog }
           : { ok: false, status: 404 },
     };
     const originalFetch = globalThis.fetch;
-    globalThis.fetch = async (input) => {
+    globalThis.fetch = (async (input: string | URL | Request) => {
       const url = String(input);
       if (url.startsWith("https://cloudflare-dns.com/dns-query")) {
         return new Response(JSON.stringify({ Status: 0 }), {
@@ -107,7 +114,7 @@ describe("SN83 CliqueAI call_subnet_surface verification (#7058)", () => {
         });
       }
       return upstreamResponse();
-    };
+    }) as typeof fetch;
     try {
       const response = await handleMcpRequest(
         new Request("https://metagraph.sh/mcp", {
@@ -126,11 +133,11 @@ describe("SN83 CliqueAI call_subnet_surface verification (#7058)", () => {
         {},
         deps,
       );
-      const result = (await response.json()).result;
+      const result = ((await response.json()) as Row).result;
       assert.equal(result.isError, false);
       assert.equal(result.structuredContent.surface_id, SURFACE_ID);
       assert.equal(result.structuredContent.status_code, 200);
-      assert.equal(result.structuredContent.body.netuid, NETUID);
+      assert.deepEqual(result.structuredContent.body, BODY);
     } finally {
       globalThis.fetch = originalFetch;
     }
