+++ tests/concentration.test.ts
@@ -8,6 +8,7 @@ import {
   loadSubnetConcentrationHistory,
   parseConcentrationHistoryWindow,
 } from "../src/concentration.ts";
+import type { Row } from "./row-type.ts";
 
 describe("computeConcentration", () => {
   test("returns null for an empty / non-array / all-zero distribution", () => {
@@ -20,13 +21,13 @@ describe("computeConcentration", () => {
 
   test("drops zero / negative / non-finite / null holders before measuring", () => {
     // Only 10, 5, and "3" (coerced) are positive holders.
-    const c = computeConcentration([10, 0, -5, Number.NaN, null, 5, "3"]);
+    const c = computeConcentration([10, 0, -5, Number.NaN, null, 5, "3"])!;
     assert.equal(c.holders, 3);
     assert.equal(c.total, 18);
   });
 
   test("a single holder is maximally concentrated (Gini 0 by definition)", () => {
-    const c = computeConcentration([42]);
+    const c = computeConcentration([42])!;
     assert.equal(c.holders, 1);
     assert.equal(c.total, 42);
     assert.equal(c.gini, 0); // one data point has no inequality
@@ -40,7 +41,7 @@ describe("computeConcentration", () => {
   });
 
   test("a perfectly uniform distribution has Gini 0 and full entropy", () => {
-    const c = computeConcentration([5, 5, 5, 5]);
+    const c = computeConcentration([5, 5, 5, 5])!;
     assert.equal(c.gini, 0);
     assert.equal(c.hhi, 0.25); // 4 × 0.25²
     assert.equal(c.hhi_normalized, 0);
@@ -50,7 +51,7 @@ describe("computeConcentration", () => {
   });
 
   test("matches hand-computed stats for [1,2,3,4]", () => {
-    const c = computeConcentration([1, 2, 3, 4]);
+    const c = computeConcentration([1, 2, 3, 4])!;
     assert.equal(c.total, 10);
     assert.equal(c.gini, 0.25);
     assert.equal(c.hhi, 0.3); // 0.1²+0.2²+0.3²+0.4²
@@ -59,12 +60,12 @@ describe("computeConcentration", () => {
     // n=4: every percentile cutoff rounds up to the top holder.
     assert.equal(c.top_1pct_share, 0.4);
     assert.equal(c.top_20pct_share, 0.4);
-    assert.ok(Math.abs(c.entropy - 1.846439) < 1e-5);
-    assert.ok(Math.abs(c.entropy_normalized - 0.923219) < 1e-5);
+    assert.ok(Math.abs(c.entropy! - 1.846439) < 1e-5);
+    assert.ok(Math.abs(c.entropy_normalized! - 0.923219) < 1e-5);
   });
 
   test("percentile cutoffs differentiate on a larger distribution", () => {
-    const c = computeConcentration([10, 9, 8, 7, 6, 5, 4, 3, 2, 1]);
+    const c = computeConcentration([10, 9, 8, 7, 6, 5, 4, 3, 2, 1])!;
     assert.equal(c.total, 55);
     assert.equal(c.gini, 0.3);
     assert.equal(c.nakamoto_coefficient, 4); // 10+9+8 = 27 ≤ 27.5; +7 > 27.5
@@ -74,20 +75,20 @@ describe("computeConcentration", () => {
   });
 
   test("a near-monopoly scores high Gini / HHI and low entropy", () => {
-    const c = computeConcentration([1000, 1, 1, 1, 1]);
-    assert.ok(c.gini > 0.7);
-    assert.ok(c.hhi > 0.9);
+    const c = computeConcentration([1000, 1, 1, 1, 1])!;
+    assert.ok(c.gini! > 0.7);
+    assert.ok(c.hhi! > 0.9);
     assert.equal(c.nakamoto_coefficient, 1);
-    assert.ok(c.entropy_normalized < 0.2);
+    assert.ok(c.entropy_normalized! < 0.2);
   });
 
   test("a sub-perfect top share does not round up to a perfect 1.0", () => {
     // Two holders, one with 99.99999% — top_1pct_share is 0.9999999, which would
     // round to a misleading 1.0 ("total concentration"). It must clamp just below
     // 1 instead, like the turnover/chain-activity ratio guards.
-    const c = computeConcentration([9_999_999, 1]);
+    const c = computeConcentration([9_999_999, 1])!;
     assert.ok(
-      c.top_1pct_share < 1,
+      c.top_1pct_share! < 1,
       `top_1pct_share must stay below 1, got ${c.top_1pct_share}`,
     );
     assert.equal(c.top_1pct_share, 0.999999);
@@ -95,7 +96,7 @@ describe("computeConcentration", () => {
 
   test("a genuine 100% share (single holder) still reports exactly 1.0", () => {
     // The clamp must only catch sub-perfect values — a true monopoly is 1.0.
-    const c = computeConcentration([500]);
+    const c = computeConcentration([500])!;
     assert.equal(c.top_1pct_share, 1);
     assert.equal(c.hhi, 1);
     assert.equal(c.hhi_normalized, 1);
@@ -113,8 +114,8 @@ describe("buildConcentration", () => {
     assert.equal(data.netuid, 7);
     assert.equal(data.neuron_count, 2);
     assert.equal(data.captured_at, "2026-06-26T02:00:00Z"); // max, not row order
-    assert.equal(data.stake.holders, 2);
-    assert.equal(data.emission.holders, 1); // the 0-emission UID is dropped
+    assert.equal(data.stake!.holders, 2);
+    assert.equal(data.emission!.holders, 1); // the 0-emission UID is dropped
   });
 
   test("cold / empty / non-array rows yield a schema-stable null block", () => {
@@ -144,12 +145,12 @@ describe("buildConcentration", () => {
     assert.equal(data.neuron_count, 3);
     assert.equal(data.entity_count, 2);
     assert.equal(data.uids_per_entity, 1.5); // 3 UIDs / 2 entities
-    assert.equal(data.stake.holders, 3); // per-UID
-    assert.equal(data.entity_stake.holders, 2); // A's two UIDs collapsed
-    assert.equal(data.entity_stake.total, 60); // A=40, B=20
+    assert.equal(data.stake!.holders, 3); // per-UID
+    assert.equal(data.entity_stake!.holders, 2); // A's two UIDs collapsed
+    assert.equal(data.entity_stake!.total, 60); // A=40, B=20
     // Validator-only: A's permitted UID (10) + B (20) — A's second UID has no permit.
-    assert.equal(data.validator_stake.holders, 2);
-    assert.equal(data.validator_stake.total, 30);
+    assert.equal(data.validator_stake!.holders, 2);
+    assert.equal(data.validator_stake!.total, 30);
   });
 
   test("the entity lens exposes concentration the per-UID lens hides", () => {
@@ -163,9 +164,9 @@ describe("buildConcentration", () => {
     assert.equal(data.neuron_count, 7);
     assert.equal(data.entity_count, 3);
     // Per-UID, W looks like 5 medium holders; per-entity, W is one ~98% whale.
-    assert.ok(data.entity_stake.gini > data.stake.gini);
-    assert.equal(data.entity_stake.nakamoto_coefficient, 1); // W alone > 50%
-    assert.ok(data.stake.nakamoto_coefficient > 1); // needs several UIDs
+    assert.ok(data.entity_stake!.gini! > data.stake!.gini!);
+    assert.equal(data.entity_stake!.nakamoto_coefficient, 1); // W alone > 50%
+    assert.ok(data.stake!.nakamoto_coefficient! > 1); // needs several UIDs
   });
 
   test("rows without a coldkey each count as their own entity", () => {
@@ -174,7 +175,7 @@ describe("buildConcentration", () => {
       1,
     );
     assert.equal(data.entity_count, 3); // none merged (missing/empty coldkey)
-    assert.equal(data.entity_stake.holders, 3);
+    assert.equal(data.entity_stake!.holders, 3);
   });
 
   test("converts D1 epoch-millisecond captured_at values to ISO strings", () => {
@@ -227,8 +228,8 @@ describe("buildConcentration", () => {
       1,
     );
     assert.equal(data.captured_at, "2026-06-26T03:00:00Z");
-    assert.equal(data.stake.holders, 1);
-    assert.equal(data.emission.holders, 1);
+    assert.equal(data.stake!.holders, 1);
+    assert.equal(data.emission!.holders, 1);
   });
 });
 
@@ -249,14 +250,14 @@ describe("parseConcentrationHistoryWindow", () => {
   });
 
   test("defaults a missing/blank window to 30d", () => {
-    assert.equal(parseConcentrationHistoryWindow(undefined).days, 30);
-    assert.equal(parseConcentrationHistoryWindow("").days, 30);
-    assert.equal(parseConcentrationHistoryWindow(null).days, 30);
+    assert.equal((parseConcentrationHistoryWindow(undefined) as Row).days, 30);
+    assert.equal((parseConcentrationHistoryWindow("") as Row).days, 30);
+    assert.equal((parseConcentrationHistoryWindow(null) as Row).days, 30);
   });
 
   test("rejects unsupported windows (incl. the longer history windows)", () => {
     for (const bad of ["1y", "all", "bogus", "0d"]) {
-      const { error } = parseConcentrationHistoryWindow(bad);
+      const { error } = parseConcentrationHistoryWindow(bad) as Row;
       assert.equal(error.parameter, "window");
       assert.match(error.message, /7d, 30d, 90d/);
     }
@@ -281,7 +282,7 @@ describe("buildConcentrationHistory", () => {
     assert.equal(data.points[1].snapshot_date, "2026-06-26");
     assert.equal(data.points[0].neuron_count, 3);
     // The newest day is concentrated (one whale); the older day is 50/50.
-    assert.ok(data.points[0].stake_gini > data.points[1].stake_gini);
+    assert.ok(data.points[0].stake_gini! > data.points[1].stake_gini!);
     assert.equal(data.points[1].stake_gini, 0);
     assert.equal(data.points[0].stake_nakamoto_coefficient, 1);
     assert.equal(typeof data.points[0].stake_top_10pct_share, "number");
@@ -320,8 +321,8 @@ describe("buildConcentrationHistory", () => {
 });
 
 describe("concentration loaders", () => {
-  function d1(rowsBySql = {}) {
-    return async (sql, _params) => {
+  function d1(rowsBySql: Record<string, Row[]> = {}) {
+    return async (sql: string, _params: unknown[]) => {
       for (const [pattern, rows] of Object.entries(rowsBySql)) {
         if (new RegExp(pattern).test(sql)) return rows;
       }
@@ -361,8 +362,8 @@ describe("concentration loaders", () => {
     );
     assert.equal(data.neuron_count, 2);
     assert.equal(data.entity_count, 1);
-    assert.equal(data.stake.holders, 2);
-    assert.equal(data.entity_stake.total, 150);
+    assert.equal(data.stake!.holders, 2);
+    assert.equal(data.entity_stake!.total, 150);
   });
 
   test("loadSubnetConcentrationHistory returns empty points on cold D1", async () => {
